requirements.txt
anthropic>=0.40.0
python-pptx>=1.0.0
Pillow>=10.0.0
pyyaml>=6.0
rich>=13.0.0
typer>=0.12.0
